+++ mrm-app/src/app/models-components/list-customers/list-customers.component.ts
@@ -5,10 +5,13 @@ import { Location } from '@angular/common';
 import { CustomerService } from '../../services/customer.service'
 import { ScriptsService } from 'src/app/services/scripts.service';
 import { BaseComponent } from 'src/app/base/base.component';
+import { MatSnackBar } from '@angular/material/snack-bar';
 
 const scripts = [
-  "../../assets/js/demo/datatables-demo.js"
+  "../../assets/js/demo/listCustomersDataTable.js"
 ]
+
+declare  var $:any;
 
 class CustomerToDisplay {
   constructor(
@@ -30,11 +33,12 @@ export class ListCustomersComponent extends BaseComponent implements OnInit {
   public constructor(
     private customerService : CustomerService,
     private activatedRoute : ActivatedRoute,
-    scriptService : ScriptsService,
+    scriptsService : ScriptsService,
     router : Router,
-    location : Location
+    location : Location,
+    matSnackBar : MatSnackBar
   ) {
-    super(scriptService, location, router)
+    super(scriptsService, location, router, matSnackBar)
    }
 
    public ngOnInit(): void {
@@ -46,17 +50,54 @@ export class ListCustomersComponent extends BaseComponent implements OnInit {
   }
 
   public ngAfterContentInit(): void {
-    this.loadScripts(scripts)
+    this.loadCustomerTableScript(scripts)
+  }
+
+  private loadCustomerTableScript(scripts) {
+    // Call the dataTables jQuery plugin
+
+    $(document).ready(function() {
+      console.log(1)
+      var listCustomersDataTable = $('#listCustomersDataTable').DataTable( {
+        "order": [[ 2, "asc" ]]
+      });
+      console.log(2)
+
+      console.log("BINDING CJSGOMET")
+      $('#listCustomersDataTable tbody').on( 'mouseenter', 'td', function () {
+          var colIdx = listCustomersDataTable.cell(this).index().column;
+          var rowIdx = listCustomersDataTable.cell(this).index().row;
+
+          $( listCustomersDataTable.cells().nodes() ).removeClass( 'highlight' );
+          $( listCustomersDataTable.rows().nodes() ).removeClass( 'highlight' );
+          $( listCustomersDataTable.column( colIdx ).nodes() ).addClass( 'highlight' );
+          $( listCustomersDataTable.row( rowIdx ).nodes() ).addClass( 'highlight' );
+      } );
+      console.log(3)
+    });
   }
 
   public updateCustomer(selectedCustomerId : number): void {
     this.router.navigate(['customers', selectedCustomerId])
   }
 
-  public deleteCustomer(selectedCustomerId : number): void {
-    this.customerService.deleteCustomer(selectedCustomerId).subscribe(
-      () => {
-        this.message = `Deleted customer!`
+  deleteCustomer(selectedCustomerId : number): void {
+    let customer = this.getCustomer(selectedCustomerId)
+    if (customer.active) {
+      customer.active = false
+      this.customerService.updateCustomer(customer).subscribe(
+        data => {
+          this.openSnackBar("Customer set to inactive.")
+          this.refreshCustomers()
+        }
+      )
+
+      return
+    }
+
+    this.customerService.deleteCustomer(customer.id).subscribe(
+      response => {
+        this.openSnackBar("Customer deleted.")
         this.refreshCustomers()
       }
     )
@@ -67,13 +108,23 @@ export class ListCustomersComponent extends BaseComponent implements OnInit {
      this.customersToDisplay
       .filter(customerToDisplay => customerToDisplay.checked)
       .forEach((selectedCustomer, index, array) => {
-        this.customerService.deleteCustomer(selectedCustomer.customer.id).subscribe(
-        () => {
-          this.message = `Deleted!`
-          if (index === array.length -1) resolve(true);
-        })
+        if (selectedCustomer.customer.active) {
+          selectedCustomer.customer.active = false
+          this.customerService.updateCustomer(selectedCustomer.customer).subscribe(
+            () => {
+              if (index === array.length -1) resolve(true);
+            }
+          )
+        } else {
+          this.customerService.deleteCustomer(selectedCustomer.customer.id).subscribe(
+            () => {
+              if (index === array.length -1) resolve(true);
+            }
+          )
+        }
       })
     }).then(() => {
+      this.openSnackBar("Customers deleted.")
       this.refreshCustomers()
     });
   }
@@ -93,6 +144,24 @@ export class ListCustomersComponent extends BaseComponent implements OnInit {
         new CustomerToDisplay(false, customer)
       )
     })
+  }
+
+  public isCustomerActive(customerId : number) {
+    let customers = this.customersToDisplay.filter((customerToDisplay => customerToDisplay.customer.id === customerId))
+    if (customers.length > 0) {
+      return customers[0].customer.active
+    }
+
+    return false
+  }
+
+  public getCustomer(customerId : number) {
+    let customers = this.customersToDisplay.filter((customerToDisplay => customerToDisplay.customer.id === customerId))
+    if (customers.length > 0) {
+      return customers[0].customer
+    }
+
+    return null
   }
 }
